Contacts: Fix issue that extra ":" at the end of number
When import from Sim Card, we can see that the imported
contact has an extra ":" at the end of the second phone
number. Since use the error split "," when parse sim
contact ANRs. Split "," is to parse sim contact emails
and split ":" is to parse sim contact ANRs.

Use the split ":" instead of "," when parse ANRs.

Change-Id: I69894d2491deac68a899fc32d21e7eb6b369f1d7
CRs-Fixed: 2216321