Contacts: Fix dialog disappears after screen rotation
During the process of deleting contacts or sharing contacts,
rotating the screen will cause the dialog to disappear. The
screen rotation will perform destruction and restart life cycle.
The dialog was not loaded and displayed when restarted.

During the process of deleting contacts, lock screen rotation.
During the process of sharing contacts, reload and show dialog
after screen rotation.

Change-Id: Iebf9b84289b6b0746a46f8e80940bbec02a2fdc5
CRs-Fixed: 2460345