Contacts: Fix pops up twice same account selection dialog
Pops up twice same account selection dialog when import contacts
to device account. The device account's name and type are "null".
When select device account and it is judged as a invalid account,
so needs user to select the account again and leads to this issue.

Add selecting device account scenario to fix this issue.

Change-Id: Ib524aff72657b5905e7640817872f19b91987bd2
CRs-Fixed: 2494912